+++ src/components/Breadcrumbs/Breadcrumb.styles.js
@@ -1,11 +1,11 @@
 import { ListUnstyled } from "../List";
 import { typography } from "../../theme";
 
-const StyledBreadcrumbList = ListUnstyled.extend`
+const StyledBreadcrumb = ListUnstyled.extend`
   font-size: ${typography.size.uno};
   font-weight: ${typography.weight.semiBold};
   display: flex;
   flex-flow: row;
 `;
 
-export default StyledBreadcrumbList;
+export default StyledBreadcrumb;
